+++ coat_bridge/ui.py
@@ -155,7 +155,6 @@ class COATBRIDGE_PT_menu(bpy.types.Panel):
 
         column.separator(factor=1.2)
         column.prop(p, "mode", text="Open as")
-        column.prop(p, "fmt", text="Format")
 
         column.separator(factor=1.2)
         grid = column.grid_flow(row_major=True, columns=2, even_columns=True, align=True)